requirements.txt
Pillow==10.3.0
numpy==1.26.4
sympy==1.12.1
cryptography==42.0.8
matplotlib==3.9.0